NOTWA - ContextException introduced - README!
When trying to add/remove BO to BOC with different context, new exception
is thrown indicating that you are trying to do something that isn't
allowed!
Source is not compilable now as the code which is not context-aware isn't
usually expecting the new exception.
Please make sure that all your code is context-aware to make the NOTWA
working again.
Please see latest commits and please utilize the ContextManager singleton
to manage contexts.