NOTWA - Security refactoring and fix
It seems that signIn method of Security class was overcomplicated and not
well formated. I have therefore fixed this rather cosmetic issue.
Moreover, there was a bug when wrong login was provided, returned user is
null. Therefore it was throwing an NullPointerException instead of
SignInException, which was also threw without message provided.