NOTWA - ParameterCollection to ParameterSet transition
It seems that ParameterCollection, as it was implemented today, was an
clumpsy approach how to implement the Collection interface. Actually,
it only brought a possibility to construct the collection by an array of
Parameters, which was no excuse for such an unnecessary lines of code.
Moreover even thought that for the inner collection the ArrayList was
choosen, the implemented interface methods behaves like Set (throwing away
duplicate entries). This was confusing and unclear!
I have therefore made a transition from implementing the specialized
Collection interface to rather extends the TreeSet (which makes the acutall
code much more compact) and renamed the class from ParameterCollection to
the ParameterSet.