NOTWA - UserDal with Security optimalization and javadoc
UserDal cooperation with Security has been optimized to faster verify
the login information. This was achived by providing a new method that
requests only the password of the given user and actually is not creating a
new User instance. This shoul avoid a chain of creating the User/Project
dependencies.

All touched sources have been javadoced.