NOTWA - TODOs solved and simple Security enhancement
At first there were several javadocs that needed to be updated - I have
coped with that.
At second there was an unneccessary duality within the Security class. The
signIn method was returning a boolean value as well as throwing an
exception to signal the actuall success of the operation. I have removed
the former so only the exception signals the wrong credentials.
Therefore the LoginDialog could be simplified.