NOTWA - Uniqe Identifier Handling
If you are creating a new BusinessObject, just not specify the uniqe
identifier. It will be pulled during the addition to the
BusinessObjectCollection. This will be only virtual id, which is not
supposed to be stored in the database. As soon as the BusinessObject
is being pushed to the database the new uniqe identifier will be acquired.