NOTWA - !README! Application package, CommandLine and DebugWindow
The NOTWA application should not be started from notwa.test but from
notwa.application, effective from now. Thus development IDE should be
adjusted to run the NOTWA from correct Class which is
notwa.application.MainClass.

As soon as you supply the /debug argument to the startup of the application
the simple DEBUG window will apear displaying the logging messages in that
orded that the last message gets to the top.

LoginDialog now contains mrneo and eter buttons to login as those guys and
the ok/storno buttons works as expected.

The default parameters of the new wit view now hides all workitems with
status closed (which is quite reasonable).

There are several small bugfixes and enhancements also which are not worthy
of even these words :-)