NOTWA - Readme - Table models refactoring
I have removed the old, (author, please forgive), a little bit awkward
table model which was overused for every table in the application
and mapping data to columns acording to the column name.

Instead of this I'm introducing a three brand new state-of-the-art table
models for each particular table.

There is also another new class ColumnSettings which defines a parameters
of single column:

    + Coumn index
    + Column header
    + Column class
    + Column alias

Every table model than holds a hashtable of these settings which allows
her (yes, table model is a woman!) to very simply achive column specific
goals, such as getColumnClass or even getValueAt. Just check by yourself!

And what about the clumsy way how to retrieve the selected workitem? Fixed
very simply by the WorkItemCollection used as a member of the WorkItemTable
class and utilizing the convertRowIndexToModel method to retrieve the
selected row correctly!

Congratulations! :-)