NOTWA - DAL - Cannot insert first note to workitem
If there was no note currently assigned to the workitem it was not possible
to create a new one. The function retrieving the note with highest uniqe
identifier returned null and we were not coping with such situation. I have
therefore changed the behavior to store the retrieved value to the temp.
variable and according to its nullness (nice word, huh?) return it or
the number 1 which is obviously the first indentifier.