NOTWA - "Verified" should not be visible
The previous fix didn't work because the SQL Parameter class traited to be
equal to only by comparsion the parameter NAME. This is wrong, because
you can define more parameters with the same name but with different value
(e.g. status == VERIFIED AND status == CLOSED).
I have therefore fixed the equals and hasCode methods to compare the
Parameter by the name and the value.