MyOzone - Fixed an issue where the XLS file won't upload due to NotSerializableException
The problem was the EJB interface, where the expected input parameter was InputStream. Streams are generally not serializable. This was originally working since the EJB was local. Now, with the remote interface, it is not possible to transfer such class over. Therefore, I have changed the interface to transfer array of bytes (as the file contents) instead.

I have also made some minor changes to the error and success reporting and rolled back one of the previous blind efforts to remove the exception.