Improvements into humidity controller
Log messages are published with the highest qos now
Startup of the script is now logged
The default of the door status changed from closed to open to prevent
unnecessary dehumidification. Otherwise, an App will notify of higher
humidity then tolerated.
The door sensor has been updated to retain the status update messages.
This way, we always get an initial value corresponding to the actual
door status.